Add _dev_config.yml to version control.
The instructions in the README now include a command that points
specifically to this file. The dev config will generally be the same
for all developers, so we might as well include it in the repository
to save new developers the confusion of trying to run the command in
its absence.